renamer components til _components for å visuelt skille mellom routes og mapper
