+++ tavla/app/(innlogget)/components/CreateBoard/NameAndFolderSelector.tsx
@@ -2,12 +2,12 @@
 import { Button, ButtonGroup } from '@entur/button'
 import { Dropdown } from '@entur/dropdown'
 import { Heading2, Label, Paragraph } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import { HiddenInput } from 'app/(innlogget)/components/Form/HiddenInput'
 import { SubmitButton } from 'app/(innlogget)/components/Form/SubmitButton'
 import { useFolderDropdown } from 'app/(innlogget)/hooks/useFolders'
 import { getFormFeedbackForField } from 'app/(innlogget)/utils/forms'
 import type { Folder } from 'app/(innlogget)/utils/types'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import rabbits from 'assets/illustrations/Rabbits.png'
 import { isNull } from 'lodash'
