+++ tavla/app/(innlogget)/components/CreateFolder/CreateFolder.tsx
@@ -3,10 +3,10 @@ import { Button, ButtonGroup, IconButton, SecondaryButton } from '@entur/button'
 import { CloseIcon, FolderIcon } from '@entur/icons'
 import { Modal } from '@entur/modal'
 import { Heading2, Label, Paragraph } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import { SubmitButton } from 'app/(innlogget)/components/Form/SubmitButton'
 import { FormError } from 'app/(innlogget)/components/FormError'
 import { getFormFeedbackForField } from 'app/(innlogget)/utils/forms'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import birds from 'assets/illustrations/Birds.png'
 import Image from 'next/image'
