+++ tavla/app/(innlogget)/components/DeleteFolder/DeleteFolder.tsx
@@ -5,6 +5,7 @@ import { CloseIcon } from '@entur/icons'
 import { Modal } from '@entur/modal'
 import { Tooltip } from '@entur/tooltip'
 import { Heading3, Paragraph, SubParagraph } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import { HiddenInput } from 'app/(innlogget)/components/Form/HiddenInput'
 import { SubmitButton } from 'app/(innlogget)/components/Form/SubmitButton'
 import { DeleteButton } from 'app/(innlogget)/oversikt/components/Column/DeleteBoard'
@@ -13,7 +14,6 @@ import {
     getFormFeedbackForField,
     type TFormFeedback,
 } from 'app/(innlogget)/utils/forms'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import ducks from 'assets/illustrations/Ducks.png'
 import Image from 'next/image'
