+++ tavla/app/(innlogget)/components/Login/Email.tsx
@@ -1,6 +1,7 @@
 'use client'
 import { ButtonGroup } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import { SubmitButton } from 'app/(innlogget)/components/Form/SubmitButton'
 import { useSearchParamsSetter } from 'app/(innlogget)/hooks/useSearchParamsSetter'
 import {
@@ -8,7 +9,6 @@ import {
     getFormFeedbackForField,
     type TFormFeedback,
 } from 'app/(innlogget)/utils/forms'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import musk from 'assets/illustrations/Musk.png'
 import { FirebaseError } from 'firebase/app'
