+++ tavla/app/(innlogget)/mapper/components/MemberAdministration/InviteUser.tsx
@@ -2,11 +2,11 @@
 import { useToast } from '@entur/alert'
 import { AddIcon } from '@entur/icons'
 import { Heading3 } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import { HiddenInput } from 'app/(innlogget)/components/Form/HiddenInput'
 import { SubmitButton } from 'app/(innlogget)/components/Form/SubmitButton'
 import { FormError } from 'app/(innlogget)/components/FormError'
 import { getFormFeedbackForField } from 'app/(innlogget)/utils/forms'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import { useActionState, useRef } from 'react'
 import type { FolderDB } from 'src/types/db-types/folders'
