+++ tavla/app/(innlogget)/oversikt/components/FoldersAndBoardsContent.tsx
@@ -1,8 +1,8 @@
 'use client'
 import { SearchIcon } from '@entur/icons'
 import { Label } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import type { Folder } from 'app/(innlogget)/utils/types'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import dynamic from 'next/dynamic'
 import { useState } from 'react'
 import type { BoardDB } from 'src/types/db-types/boards'
