+++ tavla/app/(innlogget)/tavler/[id]/rediger/components/Settings/components/WalkingDistance.tsx
@@ -1,9 +1,9 @@
 'use client'
 import { SearchableDropdown } from '@entur/dropdown'
 import { Heading4, Paragraph } from '@entur/typography'
+import ClientOnly from 'app/_components/NoSSR/ClientOnly'
 import { HiddenInput } from 'app/(innlogget)/components/Form/HiddenInput'
 import { usePointSearch } from 'app/(innlogget)/hooks/usePointSearch'
-import ClientOnly from 'app/components/NoSSR/ClientOnly'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import { useEffect, useRef } from 'react'
 import type { LocationDB } from 'src/types/db-types/boards'
