+++ tavla/app/(innlogget)/tavler/[id]/rediger/components/TileCard/components/SetOffsetDepartureTime.tsx
@@ -1,7 +1,7 @@
 import { Checkbox } from '@entur/form'
 import { Heading4, SubParagraph } from '@entur/typography'
+import ClientOnlyTextField from 'app/_components/NoSSR/TextField'
 import { TileContext } from 'app/(innlogget)/tavler/[id]/rediger/components/TileCard/context'
-import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import type { EventProps } from 'app/posthog/events'
 import {
     TRACKING_DEBOUNCE_TIME,
