+++ tavla/app/_components/Tavla123.tsx
