+++ tavla/app/_components/WordCarousel.tsx
