+++ tavla/app/demo/page.tsx
@@ -1,7 +1,7 @@
 import { Heading1, LeadParagraph } from '@entur/typography'
+import { CreateUserButton } from 'app/_components/CreateUserButton'
+import { NavigateToOversiktButton } from 'app/_components/NavigateToOversiktButton'
 import { getUserFromSessionCookie } from 'app/(innlogget)/utils/server'
-import { CreateUserButton } from 'app/components/CreateUserButton'
-import { NavigateToOversiktButton } from 'app/components/NavigateToOversiktButton'
 import type { Metadata } from 'next'
 import { DemoBoard } from './components/DemoBoard'
 import { ExpandableInformation } from './components/ExpandableInformation'
