+++ tavla/app/lag-tavle/components/CreateBoardLocally.tsx
@@ -2,12 +2,12 @@
 import { PrimaryButton } from '@entur/button'
 import { Modal } from '@entur/modal'
 import { Heading1, Heading2, Heading3, LeadParagraph } from '@entur/typography'
+import { CreateUserButton } from 'app/_components/CreateUserButton'
 import { useSaveDemoBoardInLocalStorage } from 'app/_hooks/useSaveDemoBoardInLocalStorage'
 import { TileSelector } from 'app/(innlogget)/components/TileSelector/TileSelector'
 import { formDataToTiles } from 'app/(innlogget)/components/TileSelector/utils'
 import { SettingsForm } from 'app/(innlogget)/tavler/[id]/rediger/components/Settings/components/SettingsForm'
 import { TileList } from 'app/(innlogget)/tavler/[id]/rediger/components/TileList'
-import { CreateUserButton } from 'app/components/CreateUserButton'
 import { DemoPreview } from 'app/demo/components/DemoPreview'
 import { publishBoard } from 'app/lag-tavle/actions'
 import { PublishModalContent } from 'app/lag-tavle/components/PublishBoardModal'
