+++ tavla/app/lag-tavle/components/PublishBoardModal.tsx
@@ -3,7 +3,7 @@ import { CopyableText } from '@entur/alert'
 import { PrimaryButton, SecondaryButton } from '@entur/button'
 import { LoadingDots } from '@entur/loader'
 import { Heading3, Paragraph } from '@entur/typography'
-import { CreateUserButton } from 'app/components/CreateUserButton'
+import { CreateUserButton } from 'app/_components/CreateUserButton'
 import type { PublishBoardState } from 'app/lag-tavle/components/CreateBoardLocally'
 import sheep from 'assets/illustrations/Sheep.png'
 import Image from 'next/image'
