+++ tavla/app/page.tsx
@@ -1,16 +1,16 @@
 import { Heading1, LeadParagraph } from '@entur/typography'
-import { CreateBoardButton } from 'app/components/CreateBoardButton'
+import { CreateBoardButton } from 'app/_components/CreateBoardButton'
 import type { Metadata } from 'next'
 import type { BoardDB } from 'src/types/db-types/boards'
 import { getBoardLinkServer } from 'src/utils/boardLink'
+import { FeatureShowcase } from './_components/FeatureShowcase'
+import { ImageCarousel } from './_components/ImageCarousel'
+import { NavigateToOversiktButton } from './_components/NavigateToOversiktButton'
+import { PreviewCarousel } from './_components/PreviewCarousel'
+import { Tavla123 } from './_components/Tavla123'
+import { TavlaUsageMap } from './_components/TavlaUsageMap'
+import { WordCarousel } from './_components/WordCarousel'
 import { getUserFromSessionCookie } from './(innlogget)/utils/server'
-import { FeatureShowcase } from './components/FeatureShowcase'
-import { ImageCarousel } from './components/ImageCarousel'
-import { NavigateToOversiktButton } from './components/NavigateToOversiktButton'
-import { PreviewCarousel } from './components/PreviewCarousel'
-import { Tavla123 } from './components/Tavla123'
-import { TavlaUsageMap } from './components/TavlaUsageMap'
-import { WordCarousel } from './components/WordCarousel'
 
 export const metadata: Metadata = {
     title: 'Forside | Entur Tavla - Sanntidsskjerm og avgangstavle for offentlig transport',
